Cast to int_ptr to use the correct pointer size (64 bit in case of Maix Bit)

+++ src/iostream/ostream.h
@@ -204,7 +204,7 @@ class ostream : public virtual ios {
    * \return the stream
    */
   ostream& operator<< (const void* arg) {
-    putNum(reinterpret_cast<uint32_t>(arg));
+    putNum(reinterpret_cast<intptr_t>(arg));
     return *this;
   }
   /** Output a string from flash using the Arduino F() macro.
